agents/pyproject.toml
[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[project]
name = "archipelago-agents"
version = "0.1.0"
description = "Agent implementations for Archipelago"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "aioboto3>=15.4.0",
    "aiobotocore>=2.25.0",
    "asyncpg>=0.30.0",
    "datadog-api-client>=2.44.0",
    "fastmcp>=2.12.4",
    "firecrawl-py>=4.12.0",
    "google-auth>=2.0.0",
    "httpx>=0.27.0",
    "litellm>=1.80.11",
    "loguru>=0.7.3",
    "modal>=1.0.0",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "trafilatura>=1.12.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.34.0",
    "ruff>=0.14.5",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-testmon>=2.1.0",
    "testcontainers>=4.6.1",
    "types-aioboto3>=15.2.0",
    "types-aiobotocore-sqs>=2.24.3",
]

[tool.basedpyright]
typeCheckingMode = "recommended"
reportUnusedCallResult = "none"
failOnWarnings = false
exclude = [".venv", ".git", ".tmp", "venv"]

[tool.ruff.lint]
# Enable linting rules
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "C90", # mccabe complexity
]

ignore = [
    "E501",  # Line too long
]

[tool.ruff.lint.mccabe]
max-complexity = 28
